requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.8
django-crispy-forms==1.13.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
colorama==0.4.4
requests==2.26.0
requests-oauthlib==1.3.0
django-allauth==0.46.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0